+ CoincidenceIndexAnalyzer added as independent tool
